fix(bigint): Use typeof for testing if BigInt exists
The `if (BigInt) { … }` style does not work in browser environments that do not implement BigInt (e.g. Safari <= 13).

A `hasBigInt` flag was added and exported from `util` and uses the format of `typeof BigInt !== ‘undefined’` which should work in all environments.  All usage of `BigInt` is now guarded by an equivalent of `if (hasBigInt) {…}` to ensure it’s available before use.